[dao.123] 방탈출 인증 2단계 (#79)
* feat: Auth Service/Controller 생성

* feat: Database Setting

* feat: Authentication Principal Resolver 적용

* feat: deserializer - token/member default 생성자 추가

* feat: find member controller/service

* feat: findUser By Token - TestCode

* feat: Test 후 Truncate Table 적용

* feat: login Interceptor

* feat: 예약 생성 구현

* feat: 예약 삭제 구현

* fix: exception customizing 할 수 있게 변경

* fix: Pass Reservation Test

* doc: Update Readme

* feat: database 초기화

* doc: Update Image in Readme

* refactor: reservation-schedule-theme 등록/삭제 조건 추가

* fix: 잘못 작성되어 있던 테스트 수정

* fix: Database AutoIncrement Index 설정

* refactor: token 인증 실패시 unAuthorized Exception 발생

* refactor: member 조회시 예외처리 / testcode

* refactor: schedule 조건 추가 및 테스트코드 작성 / database Auto_Increment Error 수정 - H2 DB version 변경

* refactor: Theme Testcase

* refactor: User Password encode

* refactor: jwt 설정값 property로 변경

* refactor: Message Enum 추가, 예외 처리 코드 수정

* refactor: 도메인 모델 제약조건 추가

* refactor: http-log 파일로 작성

* refactor: Login Interceptor 수정, 다른 유저의  reservation find 접근 권한 차단

* refactor: 접근제어자 수정, DAO Mapper Interface로 변경

* 잘못된 package 이름 수정

* fix: Exception Package 수정

* fix: String 값 검증 - null/empty

* fix: 반복 생성되는 객체 Injection - Singleton scope

* fix: public static bean으로 잘못 등록된 method - public bean으로 변경

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,60 +1,81 @@
 package nextstep.reservation;
 
+import nextstep.auth.JwtTokenProvider;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
+import org.springframework.dao.DuplicateKeyException;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+
+import static nextstep.auth.Interceptor.LoginInterceptor.bearer;
+import static nextstep.config.Messages.*;
+
 
 @Service
 public class ReservationService {
-    public final ReservationDao reservationDao;
-    public final ThemeDao themeDao;
-    public final ScheduleDao scheduleDao;
+    private final ReservationDao reservationDao;
+    private final ThemeDao themeDao;
+    private final ScheduleDao scheduleDao;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, JwtTokenProvider jwtTokenProvider) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
-        Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
-        if (schedule == null) {
-            throw new NullPointerException();
+    public Long create(ReservationRequest reservationRequest, String authorization) {
+        List<Schedule> schedules = scheduleDao.findById(reservationRequest.getScheduleId());
+
+        if (schedules.isEmpty()) {
+            throw new NullPointerException(NOT_FOUND_SCHEDULE.getMessage() + reservationRequest.getScheduleId());
         }
 
-        List<Reservation> reservation = reservationDao.findByScheduleId(schedule.getId());
-        if (!reservation.isEmpty()) {
-            throw new DuplicateEntityException();
+        List<Reservation> reservation = reservationDao.findByScheduleId(schedules.get(0).getId());
+        if (!reservation.isEmpty()) {  // <- 하나만 수용.. / -> size()
+            throw new DuplicateKeyException(ALREADY_REGISTERED_RESERVATION.getMessage());
         }
 
+        String accessToken = authorization.substring(bearer.length());
+        String username = jwtTokenProvider.getPrincipal(accessToken);
         Reservation newReservation = new Reservation(
-                schedule,
-                reservationRequest.getName()
+                schedules.get(0),
+                username
         );
-
         return reservationDao.save(newReservation);
     }
 
-    public List<Reservation> findAllByThemeIdAndDate(Long themeId, String date) {
-        Theme theme = themeDao.findById(themeId);
-        if (theme == null) {
+    public List<Reservation> findAllByThemeIdAndDate(Long themeId, String date, String authorization) {
+        Optional<List<Theme>> themeList = themeDao.findById(themeId);
+        if (themeList.isEmpty() || themeList.get().isEmpty()) {
             throw new NullPointerException();
         }
 
-        return reservationDao.findAllByThemeIdAndDate(themeId, date);
+        String accessToken = authorization.substring(bearer.length());
+        String username = jwtTokenProvider.getPrincipal(accessToken);
+
+        return reservationDao.findAllByThemeIdAndDate(username, themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String authorization) {
+        String accessToken = authorization.substring(bearer.length());
+        String username = jwtTokenProvider.getPrincipal(accessToken);
+
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
-            throw new NullPointerException();
+            throw new NullPointerException(RESERVATION_NOT_FOUND.getMessage());
         }
-
+       if (!Objects.equals(reservation.getName(), username)){
+            throw new AuthorizationServiceException(NOT_PERMISSION_DELETE.getMessage());
+        }
         reservationDao.deleteById(id);
     }
 }
